Consolidate the three snapshot test files into one binary
tests/snapshot_provider_test.rs, snapshot_precipitation_test.rs, and
snapshot_open_meteo_prefer_codes_test.rs each compiled as a separate
integration-test crate, so tests/helpers (wiremock setup, settings
builders) plus wiremock/tokio/insta got compiled and linked three
times over. All three are legitimate end-to-end snapshot tests driving
the real dashboard-generation pipeline (nothing misplaced) — this was
deferred pure compile-time housekeeping per
docs/test-suite-migration-plan.md Step 4.2, executed now via
docs/snapshot-test-consolidation-plan.md.

Merged into tests/snapshot_test.rs with mod provider / precipitation /
prefer_codes, test bodies ported as-is. insta derives snapshot names
from the module path, so all 16 tests/snapshots/*.snap files were
git mv'd to match (verified with `cargo insta test
--unreferenced=reject`: 16/16 pass, zero new, zero orphaned).

Build-time win (touch tests/helpers/test_utils.rs, cargo test --no-run,
isolating just the recompile+link):

|                      | before (3 binaries) | after (1 binary)  |
|----------------------|----------------------|--------------------|
| wall time            | 0.966s               | 0.779s (-19%)      |
| CPU time (user+sys)  | 7.59s                | 2.35s (-69%)       |

Wall-clock gain is modest since cargo links the old 3 binaries in
parallel across cores, but total CPU work drops ~69% — the more
representative number on CPU-constrained CI runners.
